requirements.txt
anthropic>=0.79.0
tavily-python>=0.7.21
streamlit>=1.54.0
python-dotenv>=1.2.0
postgrest>=2.28.0
langfuse>=4.0.0b1
opentelemetry-instrumentation-anthropic>=0.35.0
